Changed data's elements type of date from integer to float

+++ proyecciones/proyecciones.rb
@@ -21,7 +21,7 @@ print "#{arr}\n"
 
 #Convertir valores a enteros.
 arr.map! do |ele|
-    ele = ele.to_i
+    ele = ele.to_f
 end
 
 #Tercer test output.
